Removed incorrect usage of node width/height in MapNodeLayer
Node width/height is used by child nodes to scale their visual representations, offsets. The removed code was setting the width/height to the size of the world in pixels. This would result in all viz that uses width/height to scale proportionally to map zoom - not desirable. Instead the size is now left to be inherited from the master node (the map's viewport pixel size). Eventually the width/height of a node could be a mapped property. But if a chart is to be placed on the map, the chart will dictate width/height itself regardless.